Separated convert tests into 2 classes.
