+++ Tests/VHDLMachinesTests/VHDLMachinesConverterStatesTests.swift
@@ -12,7 +12,7 @@ import VHDLMachines
 import XCTest
 
 /// Test class for ``VHDLMachinesConverter``.
-final class VHDLMachinesConverterTests: XCTestCase {
+final class VHDLMachinesConverterStatesTests: XCTestCase {
 
     /// The machine to be converted.
     var machine = MetaMachine.initialMachine(forSemantics: .vhdl)
@@ -175,182 +175,6 @@ final class VHDLMachinesConverterTests: XCTestCase {
         ]
     }
 
-    /// The default machine attributes.
-    var machineAttributes: [AttributeGroup] {
-        [
-            AttributeGroup(
-                name: "variables",
-                fields: [
-                    Field(
-                        name: "clocks",
-                        type: .table(
-                            columns: [
-                                ("name", .line),
-                                ("frequency", .integer),
-                                ("unit", .enumerated(validValues: ["Hz", "kHz", "MHz", "GHz", "THz"]))
-                            ]
-                        )
-                    ),
-                    Field(
-                        name: "external_signals",
-                        type: .table(
-                            columns: [
-                                ("mode", .enumerated(validValues: ["in", "out", "inout", "buffer"])),
-                                ("type", .expression(language: .vhdl)),
-                                ("name", .line),
-                                ("value", .expression(language: .vhdl)),
-                                ("comment", .line)
-                            ]
-                        )
-                    ),
-                    Field(
-                        name: "generics",
-                        type: .table(
-                            columns: [
-                                ("type", .expression(language: .vhdl)),
-                                ("name", .line),
-                                ("value", .expression(language: .vhdl)),
-                                ("comment", .line)
-                            ]
-                        )
-                    ),
-                    Field(
-                        name: "machine_signals",
-                        type: .table(
-                            columns: [
-                                ("type", .expression(language: .vhdl)),
-                                ("name", .line),
-                                ("value", .expression(language: .vhdl)),
-                                ("comment", .line)
-                            ]
-                        )
-                    ),
-                    Field(
-                        name: "machine_variables",
-                        type: .table(
-                            columns: [
-                                ("type", .expression(language: .vhdl)),
-                                ("name", .line),
-                                ("value", .expression(language: .vhdl)),
-                                ("comment", .line)
-                            ]
-                        )
-                    ),
-                    Field(name: "driving_clock", type: .enumerated(validValues: ["clk"]))
-                ],
-                attributes: [
-                    "clocks": .table(
-                        [
-                            [
-                                .line("clk"),
-                                .integer(50),
-                                .enumerated("MHz", validValues: ["Hz", "kHz", "MHz", "GHz", "THz"])
-                            ]
-                        ],
-                        columns: [
-                            ("name", .line),
-                            ("frequency", .integer),
-                            ("unit", .enumerated(validValues: ["Hz", "kHz", "MHz", "GHz", "THz"]))
-                        ]
-                    ),
-                    "external_signals": .table([], columns: [
-                        ("mode", .enumerated(validValues: ["in", "out", "inout", "buffer"])),
-                        ("type", .expression(language: .vhdl)),
-                        ("name", .line),
-                        ("value", .expression(language: .vhdl)),
-                        ("comment", .line)
-                    ]),
-                    "generics": .table([], columns: [
-                        ("type", .expression(language: .vhdl)),
-                        ("name", .line),
-                        ("value", .expression(language: .vhdl)),
-                        ("comment", .line)
-                    ]),
-                    "machine_signals": .table([], columns: [
-                        ("type", .expression(language: .vhdl)),
-                        ("name", .line),
-                        ("value", .expression(language: .vhdl)),
-                        ("comment", .line)
-                    ]),
-                    "machine_variables": .table([], columns: [
-                        ("type", .expression(language: .vhdl)),
-                        ("name", .line),
-                        ("value", .expression(language: .vhdl)),
-                        ("comment", .line)
-                    ]),
-                    "driving_clock": .enumerated("clk", validValues: ["clk"])
-                ],
-                metaData: [:]
-            ),
-            AttributeGroup(
-                name: "parameters",
-                fields: [
-                    Field(name: "is_parameterised", type: .bool),
-                    Field(name: "parameter_signals", type: .table(
-                        columns: [
-                            ("type", .expression(language: .vhdl)),
-                            ("name", .line),
-                            ("value", .expression(language: .vhdl)),
-                            ("comment", .line)
-                        ]
-                    )),
-                    Field(name: "returnable_signals", type: .table(
-                        columns: [
-                            ("type", .expression(language: .vhdl)),
-                            ("name", .line),
-                            ("comment", .line)
-                        ]
-                    ))
-                ],
-                attributes: [
-                    "is_parameterised": .bool(false),
-                    "parameter_signals": .table([], columns: [
-                        ("type", .expression(language: .vhdl)),
-                        ("name", .line),
-                        ("value", .expression(language: .vhdl)),
-                        ("comment", .line)
-                    ]),
-                    "returnable_signals": .table([], columns: [
-                        ("type", .expression(language: .vhdl)),
-                        ("name", .line),
-                        ("comment", .line)
-                    ])
-                ],
-                metaData: [:]
-            ),
-            AttributeGroup(
-                name: "includes",
-                fields: [
-                    Field(name: "includes", type: .code(language: .vhdl)),
-                    Field(name: "architecture_head", type: .code(language: .vhdl)),
-                    Field(name: "architecture_body", type: .code(language: .vhdl))
-                ],
-                attributes: [
-                    "includes": .code("library IEEE;\nuse IEEE.std_logic_1164.All;", language: .vhdl),
-                    "architecture_head": .code("", language: .vhdl),
-                    "architecture_body": .code("", language: .vhdl)
-                ],
-                metaData: [:]
-            ),
-            AttributeGroup(
-                name: "settings",
-                fields: [
-                    Field(
-                        name: "initial_state", type: .enumerated(validValues: ["Initial", "Suspended", ""])
-                    ),
-                    Field(
-                        name: "suspended_state", type: .enumerated(validValues: ["Initial", "Suspended", ""])
-                    )
-                ],
-                attributes: [
-                    "initial_state": .enumerated("Initial", validValues: ["Initial", "Suspended", ""]),
-                    "suspended_state": .enumerated("Suspended", validValues: ["Initial", "Suspended", ""])
-                ],
-                metaData: [:]
-            )
-        ]
-    }
-
     /// A default VHDL machine.
     lazy var vhdlMachine = VHDLMachines.Machine(
         name: machine.name,
@@ -417,40 +241,6 @@ final class VHDLMachinesConverterTests: XCTestCase {
         let intialState = vhdlMachine.states[0]
         let attributes = converter.stateAttributes(state: intialState, machine: vhdlMachine)
         XCTAssertEqual(attributes, stateAttributes)
-    }
-
-    /// Test initial machine is set up correctly.
-    func testInitialMachine() {
-        let machine = converter.initialVHDLMachine(
-            filePath: URL(fileURLWithPath: "Test.machine", isDirectory: true)
-        )
-        let states = [
-            State(
-                name: "Initial",
-                actions: actions,
-                transitions: [],
-                attributes: stateAttributes,
-                metaData: []
-            ),
-            State(
-                name: "Suspended",
-                actions: actions,
-                transitions: [],
-                attributes: stateAttributes,
-                metaData: []
-            )
-        ]
-        let expected = MetaMachine(
-            semantics: .vhdl,
-            mutator: SchemaMutator(schema: VHDLSchema(dependencyLayout: [])),
-            name: "Test",
-            initialState: "Initial",
-            states: states,
-            dependencies: [],
-            attributes: machineAttributes,
-            metaData: []
-        )
-        XCTAssertEqual(expected, machine)
     }
 
     /// Test converter creates correct machine.
